Misc/junit5-mockito/src/main/java/workshop/Service.java
package workshop;

import java.sql.SQLException;
import java.util.Arrays;
import java.util.List;
import java.util.stream.Collectors;

/**
 * @author Subodh Kumar
 */
public class Service {
    
    private final Repository repository;

    public Service(Repository repository) {
        this.repository = repository;
    }
    
    public List<String> getStuffWithLengthLessThanFive() {
        try {
            return repository.getStuff().stream()
                    .filter(stuff -> stuff.length() < 5)
                    .collect(Collectors.toList());
        } catch (SQLException e) {
            return Arrays.asList();
        }
    }
}
